test app and compile fixes for RN android 0.29.0+
