+++ react-native/android/src/main/jni/Application.mk
@@ -12,6 +12,7 @@ APP_CPPFLAGS := -std=c++14
 APP_CPPFLAGS += -frtti
 APP_CPPFLAGS += -fexceptions
 APP_CPPFLAGS += -DREALM_HAVE_CONFIG
+APP_CPPFLAGS += -fomit-frame-pointer
 
 # Make sure every shared lib includes a .note.gnu.build-id header
 APP_LDFLAGS := -Wl,--build-id
