+++ tests/react-test-app/android/app/src/main/java/MainActivity.java
@@ -1,14 +1,6 @@
 package io.realm.react.testapp;
 
 import com.facebook.react.ReactActivity;
-import com.facebook.react.ReactPackage;
-import com.facebook.react.shell.MainReactPackage;
-
-import java.util.Arrays;
-import java.util.List;
-
-import com.rnfs.RNFSPackage;
-import io.realm.react.RealmReactPackage;
 
 public class MainActivity extends ReactActivity {
 
@@ -19,27 +11,5 @@ public class MainActivity extends ReactActivity {
     @Override
     protected String getMainComponentName() {
         return "ReactTests";
-    }
-
-    /**
-     * Returns whether dev mode should be enabled.
-     * This enables e.g. the dev menu.
-     */
-    @Override
-    protected boolean getUseDeveloperSupport() {
-        return BuildConfig.DEBUG;
-    }
-
-    /**
-     * A list of packages used by the app. If the app uses additional views
-     * or modules besides the default ones, add more packages here.
-     */
-    @Override
-    protected List<ReactPackage> getPackages() {
-        return Arrays.<ReactPackage>asList(
-            new MainReactPackage(),
-            new RNFSPackage(),
-            new RealmReactPackage()
-        );
     }
 }
